refactor: Moving events related classes and files
