+++ include/moth_ui/node.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "moth_ui/event_listener.h"
+#include "moth_ui/events/event_listener.h"
 #include "moth_ui/context.h"
 #include "moth_ui/utils/color.h"
 #include "moth_ui/blend_mode.h"
