+++ src/group.cpp
@@ -2,7 +2,7 @@
 #include "moth_ui/group.h"
 #include "moth_ui/layout/layout_entity_group.h"
 #include "moth_ui/animation/animation_clip.h"
-#include "moth_ui/event_dispatch.h"
+#include "moth_ui/events/event_dispatch.h"
 #include "moth_ui/node_clip.h"
 #include "moth_ui/context.h"
 #include "moth_ui/node_factory.h"
